before mail queue

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 class CategoryController extends Controller
@@ -21,12 +22,22 @@ class CategoryController extends Controller
 
     public function StoreCategory(Request $request)
     {
+//        dd($request->all());
         $validatedData = $request->validate([
-           'category' => 'required|string|regex:/^[a-zA-ZÑñ\s]+$/|min:3|max:35|unique:categories,category_name',
+            'category_name' => 'required|string|regex:/^[a-zA-ZÑñ\s]+$/|min:3|max:255|unique:categories,category_name',
+            'slug' => 'required|min:3|max:255|',
+            'meta_title' => 'required|min:3|max:255|',
+            'meta_description' => 'nullable|min:3|max:500|',
+            'meta_tags' => 'nullable|min:2|max:255|',
         ]);
 
         Category::create([
-            'category_name' => Str::ucfirst($request->category)
+            'category_name'        => Str::ucfirst($request->category_name),
+            'slug'                 => Str::slug($request->slug),
+            'meta_title'           => $request->meta_title,
+            'meta_description'     => $request->meta_description,
+            'meta_tags'            => $request->meta_tags,
+            'category_schema'      => json_encode($request->category_schema),
         ]);
         $notification = array(
             'message' => 'Category Created Successfully',
@@ -44,14 +55,24 @@ class CategoryController extends Controller
 
     public function UpdateCategory(Request $request, $id)
     {
-
+//dd($request->all());
         $validatedData = $request->validate([
-            'category' => 'required|string|regex:/^[a-zA-ZÑñ\s]+$/|min:3|max:35|',
+            'category_name' => 'string|regex:/^[a-zA-ZÑñ\s]+$/|min:3|max:35|',
+            'slug' => 'required|min:3|max:255|',
+            'meta_title' => 'required|min:3|max:255|',
+            'meta_description' => 'nullable|min:3|max:500|',
+            'meta_tags' => 'nullable|min:2|max:255|',
         ]);
 
-        Category::where('id', $id)->update([
-            'category_name' => Str::ucfirst($request->category)
+        Category::where('id',$id)->update([
+            'category_name'        => Str::ucfirst($request->category_name),
+            'slug'                 => Str::slug($request->slug),
+            'meta_title'           => $request->meta_title,
+            'meta_description'     => $request->meta_description,
+            'meta_tags'            => $request->meta_tags,
+            'category_schema'      => json_encode($request->category_schema),
         ]);
+
         $notification = array(
             'message' => 'Category Updated Successfully',
             'alert-type' => 'success'
@@ -59,13 +80,51 @@ class CategoryController extends Controller
         return redirect()->back()->with($notification);
     }
 
+
+
     public function DeleteCategory(Category $category)
     {
-       $category->delete();
-        $notification = array(
-            'message' => 'Category Deleted Successfully',
-            'alert-type' => 'error'
-        );
+        $verify = Post::whereRaw("FIND_IN_SET(?, category_id) > 0", $category->id)->get();
+//        dd($verify);
+        if (count($verify) > 0){
+            $multiplied = $verify->map(function ($item, $key) {
+                return explode(',', $item->category_id);
+            });
+//            dd($multiplied);
+            $multiplied2 = json_decode(json_encode($multiplied , true));
+            foreach ($multiplied2 as $key => $val){
+                if (in_array(1, $val)){
+                    $i = array_search( $category->id , $val);
+                    unset($val[$i]);
+                }else{
+                    $i = array_search( $category->id , $val);
+                    $val[$i] = 1;
+                }
+                $multiplied3[] = $val;
+            }
+
+            $collection = collect($multiplied3);
+            $convertToString = $collection->map(function ($item, $key) {
+                return implode(',', $item);
+            });
+
+            foreach ($verify as $key => $value){
+                Post::where("id", $value->id)->update(['category_id'=>$convertToString[$key]]);
+            }
+            $category->delete();
+            $notification = array(
+                'message' => 'Category Deleted Successfully',
+                'alert-type' => 'error'
+            );
         return redirect()->back()->with($notification);
+        }else{
+            $category->delete();
+            $notification = array(
+                'message' => 'Category Deleted Successfully',
+                'alert-type' => 'success'
+            );
+            return redirect()->back()->with($notification);
+        }
+//
     }
 }
